requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
jinja2==3.1.4
sentence-transformers==3.2.1
transformers==4.45.1
torch==2.4.1
numpy==1.26.4
aiohttp==3.10.10
python-dotenv==1.0.1
pytest==8.3.3
scikit-learn==1.5.2
pandas==2.2.3
streamlit==1.38.0
requests==2.32.3
plotly==5.24.1
slowapi==0.1.9
httpx==0.27.0
